refactor: move advice injector handlers into separate modules
